requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.2.3
django-debug-toolbar==5.2.0
pillow==11.3.0
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
